US-2: Order functionality extracted from Menu class
